Many of the stash values are reserved. These include: action, app, cb, controller, data, extends, format, handler, inline, json, layout, namespace, path, status, template, text and variant
Setting stash(data) interfers with the processing of render() and
causes the payload in the result to be copied from the request.
Consequently, POST status from the t-stat never sees a change flag and
doesn't request an updated config.

The stash(data) line isn't ever referenced, so just drop it.

This seems to address issue 38.