Allow specifying the loglevel with the LOGLEVEL envvar
Printing 3 lines every time the thermostat makes a request every 12
seconds in production seems a bit excessive, so allow for the
loglevel to be specified at runtime to reduce logging verbosity.
Also, increase the severity of a few serious-looking log statements
so they still get printed even when higher log levels are
specified.